+++ apps/core/src/migration/history.ts
@@ -30,6 +30,7 @@ import v9_7_6 from './version/v9.7.6'
 import v9_7_7 from './version/v9.7.7'
 import v10_0_0 from './version/v10.0.0'
 import v10_0_5 from './version/v10.0.5'
+import v10_1_0 from './version/v10.1.0'
 
 export default [
   v200Alpha1,
@@ -64,4 +65,5 @@ export default [
   v9_7_7,
   v10_0_0,
   v10_0_5,
+  v10_1_0,
 ]