+++ packages/webhook/src/types.ts
@@ -1,4 +1,5 @@
 import type { EventEmitter } from 'node:events'
+
 import type { CommentModel } from '@core/modules/comment/comment.model'
 import type { LinkModel } from '@core/modules/link/link.model'
 import type { NoteModel } from '@core/modules/note/note.model'
@@ -8,14 +9,19 @@ import type { PostModel } from '@core/modules/post/post.model'
 import type { NormalizedPost } from '@core/modules/post/post.type'
 import type { RecentlyModel } from '@core/modules/recently/recently.model'
 import type { SayModel } from '@core/modules/say/say.model'
+
 import type { ReaderModel } from '~/modules/reader/reader.model'
+
 import type { BusinessEvents } from './event.enum'
+
+export type WebhookEventSource = 'admin' | 'visitor' | 'system'
 
 export interface ExtendedEventEmitter extends EventEmitter {
   on: (<T extends BusinessEvents>(
     event: T,
     listener: (
       data: EventPayloadMapping[Extract<T, keyof EventPayloadMapping>],
+      source: WebhookEventSource,
     ) => void,
   ) => this) &
     ((event: '*', listener: (event: GenericEvent) => void) => this)